+++ src/axom/inlet/Container.cpp
@@ -209,12 +209,14 @@ Container& Container::addStructCollection(const std::string& name,
     std::vector<Key> indices;
     std::string fullName = appendPrefix(m_name, name);
     fullName = removeAllInstances(fullName, detail::COLLECTION_GROUP_NAME + "/");
-    if(m_reader.getIndices(fullName, indices))
+    const auto result = m_reader.getIndices(fullName, indices);
+    if(result == ReaderResult::Success)
     {
       container.addIndicesGroup(indices, description, true);
     }
+    markRetrievalStatus(*container.m_sidreGroup, result);
+    markAsStructCollection(*container.m_sidreGroup);
   }
-  markAsStructCollection(*container.m_sidreGroup);
   return container;
 }
 
@@ -382,9 +384,14 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<bool>(
   bool forArray,
   bool val)
 {
-  if(forArray || m_reader.getBool(lookupPath, val))
+  const auto result = m_reader.getBool(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val ? int8(1) : int8(0));
+  }
+  if(!forArray)
+  {
+    markRetrievalStatus(*sidreGroup, result);
   }
   return axom::sidre::DataTypeId::INT8_ID;
 }
@@ -396,9 +403,14 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<int>(
   bool forArray,
   int val)
 {
-  if(forArray || m_reader.getInt(lookupPath, val))
+  const auto result = m_reader.getInt(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val);
+  }
+  if(!forArray)
+  {
+    markRetrievalStatus(*sidreGroup, result);
   }
   return axom::sidre::DataTypeId::INT_ID;
 }
@@ -410,9 +422,14 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<double>(
   bool forArray,
   double val)
 {
-  if(forArray || m_reader.getDouble(lookupPath, val))
+  const auto result = m_reader.getDouble(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewScalar("value", val);
+  }
+  if(!forArray)
+  {
+    markRetrievalStatus(*sidreGroup, result);
   }
   return axom::sidre::DataTypeId::DOUBLE_ID;
 }
@@ -424,9 +441,14 @@ axom::sidre::DataTypeId Container::addPrimitiveHelper<std::string>(
   bool forArray,
   std::string val)
 {
-  if(forArray || m_reader.getString(lookupPath, val))
+  const auto result = m_reader.getString(lookupPath, val);
+  if(forArray || result == ReaderResult::Success)
   {
     sidreGroup->createViewString("value", val);
+  }
+  if(!forArray)
+  {
+    markRetrievalStatus(*sidreGroup, result);
   }
   return axom::sidre::DataTypeId::CHAR8_STR_ID;
 }
@@ -435,7 +457,7 @@ namespace detail
 {
 /*!
   *****************************************************************************
-  * \brief Adds the contents of an array to the table
+  * \brief Adds the contents of an array to the container
   * 
   * \return The keys that were added
   *****************************************************************************
@@ -455,7 +477,7 @@ std::vector<VariantKey> registerCollection(Container& container,
 
 /*!
   *****************************************************************************
-  * \brief Adds the contents of a dict to the table
+  * \brief Adds the contents of a dict to the container
   * 
   * \return The keys that were added
   *****************************************************************************
@@ -511,7 +533,8 @@ struct PrimitiveArrayHelper<Key, bool>
   {
     std::unordered_map<Key, bool> map;
     // Failure to retrieve a map is not necessarily an error
-    reader.getBoolMap(lookupPath, map);
+    const auto result = reader.getBoolMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -524,7 +547,8 @@ struct PrimitiveArrayHelper<Key, int>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, int> map;
-    reader.getIntMap(lookupPath, map);
+    const auto result = reader.getIntMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -537,7 +561,8 @@ struct PrimitiveArrayHelper<Key, double>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, double> map;
-    reader.getDoubleMap(lookupPath, map);
+    const auto result = reader.getDoubleMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -550,7 +575,8 @@ struct PrimitiveArrayHelper<Key, std::string>
                                      const std::string& lookupPath)
   {
     std::unordered_map<Key, std::string> map;
-    reader.getStringMap(lookupPath, map);
+    const auto result = reader.getStringMap(lookupPath, map);
+    markRetrievalStatus(*container.sidreGroup(), result);
     return registerCollection(container, map);
   }
 };
@@ -849,41 +875,48 @@ Container& Container::registerVerifier(std::function<bool(const Container&)> lam
 
 bool Container::verify() const
 {
-  bool verified = true;
+  // Whether the calling container has anything in it
+  // If the name is empty then we're the global (root) container, which we always
+  // consider to be defined
+  const bool this_container_defined = isUserProvided() || m_name.empty();
+
   // If this container was required, make sure something was defined in it
-  verified &=
-    verifyRequired(*m_sidreGroup, static_cast<bool>(*this), "Container");
+  bool verified =
+    verifyRequired(*m_sidreGroup, this_container_defined, "Container");
+
   // Verify this Container if a lambda was configured
-  if(m_verifier && !m_verifier(*this))
+  if(this_container_defined && m_verifier && !m_verifier(*this))
   {
     verified = false;
     SLIC_WARNING(
       fmt::format("[Inlet] Container failed verification: {0}", m_name));
   }
-  // Verify the child Fields of this Container
-  for(const auto& field : m_fieldChildren)
-  {
-    if(!field.second->verify())
-    {
-      verified = false;
-    }
-  }
-  // Verify the child Containers of this Container
-  for(const auto& container : m_containerChildren)
-  {
-    if(!container.second->verify())
-    {
-      verified = false;
-    }
-  }
 
-  // Verify the child Functions of this Container
-  for(const auto& function : m_functionChildren)
+  // Checking the child objects is not needed if the container wasn't defined in the input file
+  if(this_container_defined)
   {
-    if(!function.second->verify())
+    // Verify the child Fields of this Container
+    for(const auto& field : m_fieldChildren)
     {
-      verified = false;
+      verified = verified && field.second->verify();
     }
+    // Verify the child Containers of this Container
+    for(const auto& container : m_containerChildren)
+    {
+      verified = verified && container.second->verify();
+    }
+
+    // Verify the child Functions of this Container
+    for(const auto& function : m_functionChildren)
+    {
+      verified = verified && function.second->verify();
+    }
+  }
+  // If this has a collection group, it always needs to be verified, as annotations
+  // may have been applied to the collection group and not the calling group
+  else if(hasContainer(detail::COLLECTION_GROUP_NAME))
+  {
+    verified = verified && getContainer(detail::COLLECTION_GROUP_NAME).verify();
   }
 
   return verified;
@@ -1000,13 +1033,13 @@ bool Container::contains(const std::string& name) const
 {
   if(auto container = getChildInternal<Container>(name))
   {
-    // call operator bool on the container itself
-    return static_cast<bool>(*container);
+    // Check if the container itself exists
+    return container->exists();
   }
   else if(auto field = getChildInternal<Field>(name))
   {
-    // call operator bool on the field itself
-    return static_cast<bool>(*field);
+    // Check if the field itself exists
+    return field->exists();
   }
   else if(auto function = getChildInternal<Function>(name))
   {
@@ -1016,23 +1049,52 @@ bool Container::contains(const std::string& name) const
   return false;
 }
 
-Container::operator bool() const
+bool Container::exists() const
 {
-  // Check if any of its child containers are nontrivial
+  // Check if any of its child containers exist
   const bool has_containers =
     std::any_of(m_containerChildren.begin(),
                 m_containerChildren.end(),
                 [](const decltype(m_containerChildren)::value_type& entry) {
-                  return static_cast<bool>(*entry.second);
+                  return entry.second->exists();
                 });
 
   const bool has_fields =
     std::any_of(m_fieldChildren.begin(),
                 m_fieldChildren.end(),
                 [](const decltype(m_fieldChildren)::value_type& entry) {
+                  return entry.second->exists();
+                });
+
+  // Functions cannot be defaulted and thus have an unambiguous operator bool
+  const bool has_functions =
+    std::any_of(m_functionChildren.begin(),
+                m_functionChildren.end(),
+                [](const decltype(m_functionChildren)::value_type& entry) {
                   return static_cast<bool>(*entry.second);
                 });
 
+  return has_containers || has_fields || has_functions;
+}
+
+bool Container::isUserProvided() const
+{
+  // Check if any of its child containers had user-provided fields
+  const bool has_containers =
+    std::any_of(m_containerChildren.begin(),
+                m_containerChildren.end(),
+                [](const decltype(m_containerChildren)::value_type& entry) {
+                  return entry.second->isUserProvided();
+                });
+
+  const bool has_fields =
+    std::any_of(m_fieldChildren.begin(),
+                m_fieldChildren.end(),
+                [](const decltype(m_fieldChildren)::value_type& entry) {
+                  return entry.second->isUserProvided();
+                });
+
+  // Functions cannot be defaulted and thus have an unambiguous operator bool
   const bool has_functions =
     std::any_of(m_functionChildren.begin(),
                 m_functionChildren.end(),